% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{cleanjoinweed}
\alias{cleanjoinweed}
\title{Merge the Datavant word list and the PHI word list, remove whitelist}
\usage{
cleanjoinweed(
  phivec,
  datavantvec,
  whitevec = character(0),
  verbose = 0,
  split = TRUE
)
}
\arguments{
\item{phivec}{vector of terms from user (usually)}

\item{datavantvec}{vector of terms from datavant (usually)}

\item{whitevec}{vector of terms to exclude from the new list. Default is empty.}

\item{verbose}{How much should be printed to the console? 0 (default) for none.  Higher values potentially display more}

\item{split}{Should terms with more than one word (space or hyphen) be split and subterms included? Default is TRUE}
}
\value{
A data.frame suitable for saving as a Scrubber configuration file, invisibly
}
\description{
Merge the Datavant word list and the PHI word list, remove whitelist
}
