% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Helper file for master(). Not expected for direct call by user.}
\usage{
expand(
  namesforblacklist,
  originalconfigdir,
  temporaryconfigdir,
  verbose = 0,
  overwriteifexists = FALSE
)
}
\arguments{
\item{namesforblacklist}{Location of the PHI file.  csv file that contains first names, middle names, last names, cities, and counties}

\item{originalconfigdir}{Location of the Datavant scrubber configuration directory}

\item{temporaryconfigdir}{Location to store expanded databases adapted to the file to be scrubbed.}

\item{verbose}{How much should be printed to the console? 0 (default) for none.  Higher values potentially display more}

\item{overwriteifexists}{If the expanded databases already exist, should they be overwritten}
}
\value{
invisible write.table()
}
\description{
Helper file for master(). Not expected for direct call by user.
}
