% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{readoriginal}
\alias{readoriginal}
\title{Function to read Datavant databases}
\usage{
readoriginal(filebase, inconfigdir, verbose = 0, quiet = FALSE)
}
\arguments{
\item{filebase}{which database? e.g., "firstname_db"}

\item{inconfigdir}{Directory with existing datavant configuration files}

\item{verbose}{How much should be printed to the console? 0 (default) for none.  Higher values potentially display more}

\item{quiet}{Suppress warning messages while reading?  Default = FALSE}
}
\value{
invisible data.frame
}
\description{
Function to read Datavant databases
}
