% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removemrn.R
\name{removemrn}
\alias{removemrn}
\title{Use gsub() to remove strings of numbers that may be MRNs. 8 or 9 digit MRN can be specified.}
\usage{
removemrn(
  dirwithmrn = ".",
  dirwithoutmrn = ".",
  verbose = 0,
  digits_in_mrn = 9,
  flag = "XXX-MRNMRN",
  perllike = TRUE,
  case_sensitive = TRUE,
  listofcolumnnamestoscrub = list(OPH_ORDER_text = c("ORDER_COMMENT", "NARRATIVE"),
    OPH_SURGERY_text = c("OP_NOTE", "BRIEF_OP_NOTE"), OPH_ENC_PROBLEM_LIST_text =
    c("PROBLEM_CMT"), OPH_ENC_HPI_text = c("HPI_COMMENT"), OPH_ENC_VISIT_SUMMARY_text =
    c("HPI", "TECH_COMMENT", "PAT_OCULAR_HISTORY"), OPH_LAB_text = c("LINE_COMMENT",
    "RESULTS_COMP_CMT", "RESULTS_CMT"), OPH_ENC_VISIT_text = c("PROGRESS_NOTE"),
    OPH_RAD_text = c("APPT_NOTE", "TEXT"), OPH_FAMILY_HX_text = c("COMMENTS",
    "FAM_STAT_NAME"), OPH_ENC_EXAM_text = c("SMRTDTA_ELEM_VALUE"), 
    
    OPH_LAB_ORDER_text = c("SPECIMEN_COMMENTS"), OPH_MED_text = c("INSTRUCTION"),
    OPH_SURGERY_all_text = c("PRE_OP_DIAG"), OPH_SURGERY_PROCEDURE_text = c("COMMENTS"))
)
}
\arguments{
\item{dirwithmrn}{location/folder/directory of files with MRNs.}

\item{dirwithoutmrn}{location/folder/directory to write files without MRNs.}

\item{verbose}{Amount of output provided to console. 0 (default) for none.  Higher values may provide more.}

\item{digits_in_mrn}{Number of digits in MRN. Default is 9.}

\item{flag}{Character string to replace found MRNs. Default is "XXX-MRNMRN".}

\item{perllike}{Should perl regex engine (PCRE) be used (TRUE, default) or TRE (FALSE).}

\item{case_sensitive}{Should variable names be found sensitive to case (TRUE, default) or insensitive to case (FALSE).}

\item{listofcolumnnamestoscrub}{which columns in which files to scrub? This argument is a list. Each component is a vector of column names.  The component names are the unique parts of the file names of the files to scrub.  Default value is to include 22 variables from 14 files.}
}
\value{
invisible NULL. But several files are created as a side effect.
}
\description{
Use gsub() to remove strings of numbers that may be MRNs. 8 or 9 digit MRN can be specified.
}
