% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{writenew}
\alias{writenew}
\title{Function to write possibly augmented databases}
\usage{
writenew(df, filebase, outconfigdir, overwriteifexists = FALSE)
}
\arguments{
\item{df}{data.frame to write}

\item{filebase}{which database? e.g., "firstname_db"}

\item{outconfigdir}{directory to write data}

\item{overwriteifexists}{Overwrite existing configuration file if it exists? Default is FALSE}
}
\value{
invisible NULL
}
\description{
Function to write possibly augmented databases
}
